fix: function isVisit
